fix: move EditorView outside of plugins
