+++ packages/picasso-rich-text-editor/src/LexicalEditorView/index.ts
@@ -1,2 +1,2 @@
-export { default } from './EditorView'
+export { default } from './LexicalEditorView'
 export { default as lexicalEditorStyles } from './styles'
