+++ packages/picasso-rich-text-editor/src/LexicalEditorView/styles.ts
