+++ packages/picasso-rich-text-editor/src/QuillEditor/LazyQuillEditor.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef, lazy, Suspense } from 'react'
 
-import QuillEditorView from '../plugins/EditorView'
+import QuillEditorView from '../LexicalEditorView'
 import type { Props } from './QuillEditor'
 
 const QuillEditor = lazy(() => import('./QuillEditor'))
